migrate assertion utils
