+++ plugins/cost-insights/src/components/CostOverviewCard/CostOverviewCard.tsx
@@ -26,11 +26,11 @@ import { useScroll, useFilters, useConfig } from '../../hooks';
 import { mapFiltersToProps } from './selector';
 import { DefaultNavigation } from '../../utils/navigation';
 import { formatPercent } from '../../utils/formatters';
+import { findAlways } from '../../utils/assert';
 import {
   Cost,
   CostInsightsTheme,
   MetricData,
-  findAlways,
   getComparedChange,
 } from '../../types';
