+++ plugins/cost-insights/src/components/ResourceGrowthBarChartLegend/ResourceGrowthBarChartLegend.test.tsx
@@ -17,7 +17,9 @@
 import React, { PropsWithChildren } from 'react';
 import { renderInTestApp } from '@backstage/test-utils';
 import { ResourceGrowthBarChartLegend } from './ResourceGrowthBarChartLegend';
-import { defaultCurrencies, Duration, findAlways } from '../../types';
+import { Duration } from '../../types';
+import { defaultCurrencies } from '../../utils/currency';
+import { findAlways } from '../../utils/assert';
 import { MockConfigProvider, MockCurrencyProvider } from '../../utils/tests';
 
 const engineers = findAlways(defaultCurrencies, c => c.kind === null);
